*Large restructuring of the SLURM submission scripts
Now running the correction on all anat directories in the parent directory following the sub-XXX/ses-XX/anat path. Now it is also possible to mandatory an output directory and the working directory are created inside this output directory if the -w flag is set. Otherwise it is created in tmp and deleted at the end of the script