*Add multiple features to MagnOnly GNLC: Jacobian modulation, rename and move results to output_dir, copy json files, remove working_dir // Remove finished TODOs from README
Add flags for all the new functions: no-jacobian (run jacobian or not after GNLC) and delete-workdir (delete or preserve working directory) // dependency for jobs (sequential processing) // Previous changes that were still as TODOs in README: 1. subjects and sessions specification in the CLI command (flags: -sub and -ses), 2. check if data is already present in output directory (if so, skip this session)