Check if NonlinearGradientCorrection key in sidecar json is already true + adjust JSON metadata after correction
For both MagOnly and RealImag correction, the metadata in the sidecar JSON is checked and the processing is aborted if the data is already gradient nonlinearity corrected. If false, the processing jobs are submitted to the SLURM cluster. After the correction the sidecar JSON of each corrected NIfTI file is adjusted by setting the NonlinearGradientCorrection to true and the NonlinearGradientCorrectionType to 3D.